STM32WB Enable BLE HCIDriver for all targets
+ astyle
